modification src/BmiService.java

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double heightMetre, double weightKg){
-        return (int)(weightKg/Math.pow(heightMetre, 2));
+    public double calculate(double heightMetre, double weightKg) {
+        return weightKg / Math.pow(heightMetre, 2);
     }
 }
